BIG MOVES
Got the plotting on the fly flow working for all of our plot types, figured out how to link up the cut points slider with the plot (not yet dynamic with data but used dummy values for the test dataset). Added checkboxes for showing the centerpoint and adding labels (this doesn't work yet, ggplot aes length error). Lots more to do but this is getting very cool!