First reactive function is working!
Starting to figure out how to chunk things out and remove them from the long RenderPlot section; very basic but I got "dataframefinder" working to choose the dataframe from the upload or default. This gives output that can be displayed; eventually we can have this working to show the center point, etc.